requirements.txt
django
djangorestframework
channels
daphne
paho-mqtt
requests
asgiref
twisted
autobahn
txaio
cryptography
pyopenssl
service-identity
attrs
msgpack
cbor2
ujson
py-ubjson
